Fixed ImportError for '_TextTestResult' on Python 3.12; fixed 'addDuration' warning
The `unittest._TextTestResult` alias was removed in Python 3.12 in favor
of the `unittest.TextTestResult` class.  This class is overloaded in `result.py`
to implement some additional error classes and hooks.

Fixed by trying another import location, and using a local alias for the base class.
https://docs.python.org/3/whatsnew/3.12.html#id3

At invocation time, the test runner would also periodically emit warnings

    warnings.warn("TestResult has no addDuration method...

This appears to be a new Python 3.12 feature to track duration of tests for the
overall report, which isn't included in our subclass.  Added a stub to satisfy
this interface and silence the warnings.

https://docs.python.org/3/whatsnew/3.12.html#unittest
https://github.com/nose-devs/nose2/issues/569

Bumped package version to '1.3.7.3'